rename prompt_template to template
