+++ src/dbally/iql_generator/iql_prompt_template.py
@@ -3,7 +3,7 @@ from typing import List
 from dbally.exceptions import DbAllyError
 from dbally.prompts import PromptTemplate
 from dbally.prompts.elements import FewShotExample
-from dbally.prompts.prompt_template import PromptFormat
+from dbally.prompts.template import PromptFormat
 from dbally.views.exposed_functions import ExposedFunction
