+++ src/dbally/prompts/__init__.py
@@ -1,3 +1,3 @@
-from .prompt_template import ChatFormat, PromptTemplate, PromptTemplateError
+from .template import ChatFormat, PromptTemplate, PromptTemplateError
 
 __all__ = ["PromptTemplate", "PromptTemplateError", "ChatFormat"]
