+++ src/dbally/view_selection/view_selector_prompt_template.py
@@ -2,7 +2,7 @@ from typing import Dict, List
 
 from dbally.prompts import PromptTemplate
 from dbally.prompts.elements import FewShotExample
-from dbally.prompts.prompt_template import PromptFormat
+from dbally.prompts.template import PromptFormat
 
 
 class ViewSelectionPromptFormat(PromptFormat):
